% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_output.R
\name{delete_output}
\alias{delete_output}
\title{Delete files from output folder}
\usage{
delete_output(folder = "output", extensions = c("xlsx", "xls"), which = NULL)
}
\arguments{
\item{which}{}
}
\description{
Delete files from output folder
}
