% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_safely.R
\name{export_safely}
\alias{export_safely}
\title{Export files to the output folder}
\usage{
export_safely(
  data,
  folder = "output/data",
  name = NULL,
  format = "xlsx",
  overwrite = FALSE,
  time_in_name = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Export files to the output folder
}
