% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_safe_workspace_name.R
\name{get_safe_workspace_name}
\alias{get_safe_workspace_name}
\title{Get a safe name to save current workspace without overwriting}
\usage{
get_safe_workspace_name(name = "analysis", time_in_name = TRUE)
}
\arguments{
\item{name}{A desired name for the saved workspace. If no name is provided,
the name will be 'analysis'.}

\item{time_in_name}{A logical to indicate whether a timestamp
should be included in the workspace's name.}
}
\value{
A safe name for exporting the workspace, as a "character string", and also indicated in a message.
}
\description{
Get a safe name to save current workspace without overwriting
}
\examples{
\donttest{
if (FALSE) {
  library(phdcocktail)
  get_safe_workspace_name()
}
}

}
