% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibd_data1}
\alias{ibd_data1}
\alias{ibd_data2}
\title{Inflammatory Bowel Disease (IBD) datasets}
\format{
Two data frames with each 30 rows and six columns:
\describe{
\item{patientid}{Patient ID}
\item{gender}{Gender}
\item{disease_location}{Disease location}
\item{disease_behaviour}{Disease behaviour}
\item{crp_mg_l}{C-reactive protein (mg/L)}
\item{calprotectin_ug_g}{Faecal calprotectin (ug/g)}
}
}
\source{
Randomly generated data
}
\usage{
ibd_data1

ibd_data2
}
\description{
'ibd_data1' and 'ibd_data2' are two small datasets containing data collected from IBD patients,
more specifically patients with Crohn's disease. 'ibd_data2' is a modified version of 'ibd_data1'
by introducing missing and incorrect entries 'L11' into the column 'disease_location'.
}
\keyword{datasets}
