% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibd_data_dict}
\alias{ibd_data_dict}
\title{Data dictionary for Inflammatory Bowel Disease (IBD) data}
\format{
A data frame with 53 rows and four columns:
\describe{
\item{variable}{Variable name in the 'short', i.e. 'excel', form}
\item{variable_label}{Variable name in the publication form}
\item{value}{Value name in the 'short', i.e. 'excel', form}
\item{value_label}{Value name in the publication form}
}
}
\usage{
ibd_data_dict
}
\description{
A small, non-exhaustive list of variables that are commonly collected in IBD research. For each
variable and its levels, if applicable, publications-ready labels are provided
}
\keyword{datasets}
