% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibd_outcomes}
\alias{ibd_outcomes}
\title{Inflammatory Bowel Disease (IBD) outcomes}
\format{
A data frame with eight rows and seven columns:
\describe{
\item{outcome}{Outcome type}
\item{timepoint}{Assessment timepoint}
\item{achieved}{Number of patients who achieved the outcome}
\item{total}{Total number of patients}
\item{proportion}{Proportion of patients who achieved the outcome}
\item{percentage}{Percentage of patients who achieved the outcome}
\item{percentage_labelled}{Percentage of patients who achieved the outcome, suffixed with '\%'}
}
}
\usage{
ibd_outcomes
}
\description{
A table containing proportions and percentages of IBD patients achieving clinical outcomes.
}
\keyword{datasets}
