% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_recent_work.R
\name{identify_recent_work}
\alias{identify_recent_work}
\title{Identify the most recent saved R workspace}
\usage{
identify_recent_work(folder = "output")
}
\arguments{
\item{folder}{The folder in which the workspace need to be identified.}
}
\value{
A message indicating the most recent saved workspace.
}
\description{
Identify the most recent saved R workspace
}
\examples{
\donttest{
library(phdcocktail)
if (FALSE) {
  identify_recent_work()
}
}

}
