% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_recent_workspace.R
\name{identify_recent_workspace}
\alias{identify_recent_workspace}
\title{Identify the most recent saved R workspace}
\usage{
identify_recent_workspace(folder = "output")
}
\arguments{
\item{folder}{The folder in which the workspace need to be identified.}
}
\value{
The most recent saved workspace, as a "character string", and also indicated in a message.
}
\description{
Identify the most recent saved R workspace
}
\examples{
\donttest{
library(phdcocktail)
if (FALSE) {
  identify_recent_workspace()
}
}

}
