% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot \% of outcomes as bars}
\usage{
plot_bars(
  data,
  outcome,
  proportion,
  percentage_labelled,
  achieved,
  total,
  x_axis_title = NULL,
  y_axis_title = "\% Patients",
  legend_title = "Outcome",
  bar_fill = "Greys",
  grouping = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot \% of outcomes as bars
}
