% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_bars.R
\name{plot_outcome_bars}
\alias{plot_outcome_bars}
\title{Plot \% of outcomes as bars}
\usage{
plot_outcome_bars(
  data,
  outcome_columns,
  bar_order = NULL,
  outcome_names = NULL,
  x_axis_name = NULL,
  y_axis_name = "\% Patients",
  legend_title = "Outcome",
  na_to_zero = TRUE,
  bar_fill = "Greys",
  show_numbers = "all",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot \% of outcomes as bars
}
