% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quantiles_report.R
\name{print.quantiles_report}
\alias{print.quantiles_report}
\title{Define a custom print method for the quantiles_report class and make it a global function}
\usage{
\method{print}{quantiles_report}(summary_data)
}
\arguments{
\item{summary_data}{}
}
\description{
Define a custom print method for the quantiles_report class and make it a global function
}
