% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vrs.R
\name{recode_vrs}
\alias{recode_vrs}
\title{Recode specific variables based on data dictionary}
\usage{
recode_vrs(data, data_dictionary, vrs, factor = FALSE)
}
\arguments{
\item{data}{input data set}

\item{data_dictionary}{name of the data dictionary}

\item{vrs}{a vector with the variables to be recoded}

\item{factor}{alogical to indicate whether the recoded variables need to be converted to factors}
}
\value{
the input data set with recoded and labelled variables
}
\description{
Recode specific variables based on data dictionary
}
