% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_quantiles.R
\name{report_quantiles}
\alias{report_quantiles}
\title{Report median-quantiles summaries}
\usage{
report_quantiles(data, summary_vrs, grouping_vrs = NULL)
}
\arguments{
\item{grouping_vrs}{}
}
\description{
Report median-quantiles summaries
}
