% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_work.R
\name{save_work}
\alias{save_work}
\title{Save global objects to the output folder}
\usage{
save_work(
  folder = "output/Rdata",
  name = "analysis",
  objects_to_save = NULL,
  time_in_name = TRUE
)
}
\arguments{
\item{time_in_name}{}
}
\description{
Save global objects to the output folder
}
