% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_fresh.R
\name{start_fresh}
\alias{start_fresh}
\title{Restart R session}
\usage{
start_fresh()
}
\value{
A clean R session
}
\description{
Restart R session
}
\examples{
\donttest{
if (FALSE) {
library(phdcocktail)
start_fresh()
  }
}

}
