% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_output.R
\name{delete_output}
\alias{delete_output}
\title{Delete files from the 'output' folder}
\usage{
delete_output(folder = "output", which = NULL, extensions = c("xlsx", "xls"))
}
\arguments{
\item{folder}{A folder from which the user wishes to delete files.
The default folder is 'output'.}

\item{which}{An argument through which the user can specify which file to be deleted.
There are 3 valid character values for this argument: i) 'except-r' option, which will delete
all files except for R script and data files, ii) 'all' option, which will delete all files in
the specified folder, iii) 'extensions' option, which specify the extensions to be deleted.
The default behaviour is 'except-r'.}

\item{extensions}{A character vector indicating extensions of files that the user wishes to
delete. The default extensions are excel extensions.}
}
\value{
A message listing the files that have been deleted.
}
\description{
Delete files from the 'output' folder
}
\examples{
\dontrun{
delete_output()
}

}
