% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_safely.R
\name{export_safely}
\alias{export_safely}
\title{Export files to the 'output' folder}
\usage{
export_safely(
  data,
  folder = "output/data",
  name = NULL,
  format = "xlsx",
  overwrite = FALSE,
  time_in_name = FALSE,
  ...
)
}
\arguments{
\item{data}{The object to be exported.}

\item{folder}{The folder to which the object need to be exported.
The default is a 'data' folder within the 'output' folder.}

\item{name}{The name of the exported file. If no name is provided,
the file will inherit the object's name.}

\item{format}{The format of the exported file. Default is 'xlsx'.}

\item{overwrite}{A logical to indicate whether preexisting files
with identical names should be overwritten. Default is 'FALSE'.}

\item{time_in_name}{A logical to indicate whether a timestamp
should be included in the file's name.}

\item{...}{Other argument that can be passed to 'rio::export'}
}
\value{
A message indicating the full path of the exported file.
}
\description{
Export files to the 'output' folder
}
\examples{
\dontrun{
library(phdcocktail)
export_safely(mtcars)
}

}
