% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quantiles_report.R
\name{print.quantiles_report}
\alias{print.quantiles_report}
\title{A custom print method for the 'quantiles_report' class}
\usage{
\method{print}{quantiles_report}(x, ...)
}
\arguments{
\item{x}{A data frame of the class 'quantiles_report'.}

\item{...}{Other argument that can be passed to 'print'.}
}
\value{
The function displays the content of the column 'report'
in separate lines.
}
\description{
A custom print method for the 'quantiles_report' class
}
\examples{
\donttest{
if (FALSE) {
library(phdcocktail)
summary_data <- report_quantiles(mtcars, summary_vrs = "mpg")
print(summary_data)
  }
}

}
