% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vrs.R
\name{recode_vrs}
\alias{recode_vrs}
\title{Recode variables and their levels based on a data dictionary}
\usage{
recode_vrs(data, data_dictionary, vrs, factor = FALSE)
}
\arguments{
\item{data}{A data frame with raw data.}

\item{data_dictionary}{A data dictionary containing labels for the variables and their levels.}

\item{vrs}{A character vector specifying the variables to be recoded.}

\item{factor}{A logical to indicate whether recoded variables need to be converted to factors.}
}
\value{
The input data frame with recoded and labelled variables.
}
\description{
Recode variables and their levels based on a data dictionary
}
\examples{
try(ibd_data_recoded <- recode_vrs(data = ibd_data, data_dictionary = ibd_data_dict,
vrs = c("disease_location", "disease_behaviour", "gender"), factor = TRUE))

}
