% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_quantiles.R
\name{report_quantiles}
\alias{report_quantiles}
\title{Report median-quantiles summaries}
\usage{
report_quantiles(data, summary_vrs, grouping_vrs = NULL)
}
\arguments{
\item{data}{A data frame including numeric variables to be summarized.}

\item{summary_vrs}{A character vector specifying the numeric variables to be summarized.}

\item{grouping_vrs}{A character vector specifying the grouping variables, if any.}
}
\value{
A dataframe of the class 'quantiles_report', containing a
'report' column, which report the 'median (quartile 1-quartile 3)'
combinations for each specified numeric variable, at each grouping
key.
}
\description{
Report median-quantiles summaries
}
\examples{
\donttest{
if (FALSE) {
library(phdcocktail)
summary_data <- report_quantiles(mtcars, summary_vrs = "mpg")
print(summary_data)
  }
}

}
