% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_work.R
\name{save_work}
\alias{save_work}
\title{Save (some) global objects without overwriting}
\usage{
save_work(name = "analysis", objects_to_save = NULL, time_in_name = TRUE)
}
\arguments{
\item{name}{The name of the saved workspace. If no name is provided,
the name will be 'analysis'.}

\item{objects_to_save}{A vector specifying which objects from the
workspace need to be saved. The default is 'NULL', which will save
all objects in the global environment.}

\item{time_in_name}{A logical to indicate whether a timestamp
should be included in the file's name.}
}
\value{
A message indicating full paths to the saved '.Rdata'
file and a '.txt' list of its objects.
}
\description{
Save (some) global objects without overwriting
}
\examples{
\donttest{
if (FALSE) {
  library(phdcocktail)
  save_work()
}
}

}
